Google Maps with Points!
I've managed to plot all the points from January's dataset! We can swap
out the URL to a different Socrata view to see different datasets.

We should look at how we want to visually present it now. Probably a
Bootstrap layout, maybe some buttons to swap the dataset we are looking
at.

Note: This files name is different than the current live map, as of this
writting.